requirements.txt
Pillow
requests
selenium
pytz
beautifulsoup4
pytesseract
opencv-python
discord.py